[ui-core] Upgrade js/ts linters to the latest possible versions
White space has been adjusted in some files as the default linting rules have changed with the newer versions.

Updated Vue linting rules have been ignored given that new development will be in React only.
