+++ desktop/core/src/desktop/js/parse/sqlStatementsParser.d.ts
@@ -1,0 +1,10 @@
+import { ParsedLocation } from 'parse/types';
+
+export interface ParsedSqlStatement {
+  firstToken: string;
+  statement: string;
+  location: ParsedLocation;
+  type: string;
+}
+
+export function parse(statement: string): ParsedSqlStatement[];
