+++ package.json
@@ -102,8 +102,8 @@
     "@types/react-dom": "18.0.6",
     "@types/sanitize-html": "1.27.0",
     "@types/webpack": "5.28.0",
-    "@typescript-eslint/eslint-plugin": "4.25.0",
-    "@typescript-eslint/parser": "4.25.0",
+    "@typescript-eslint/eslint-plugin": "7.16.0",
+    "@typescript-eslint/parser": "7.16.0",
     "@vue/compiler-sfc": "3.2.0",
     "@vue/server-renderer": "3.2.0",
     "@vue/test-utils": "2.4.6",
@@ -115,11 +115,11 @@
     "clean-webpack-plugin": "4.0.0",
     "copy-webpack-plugin": "12.0.2",
     "css-loader": "5.2.6",
-    "eslint": "7.27.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-jest": "25.2.4",
-    "eslint-plugin-prettier": "3.4.0",
-    "eslint-plugin-vue": "7.10.0",
+    "eslint": "8.56.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-jest": "28.6.0",
+    "eslint-plugin-prettier": "5.1.3",
+    "eslint-plugin-vue": "9.27.0",
     "expose-loader": "3.0.0",
     "grunt": "1.5.3",
     "grunt-contrib-less": "3.0.0",
@@ -134,7 +134,7 @@
     "load-grunt-tasks": "5.1.0",
     "postcss-less": "6.0.0",
     "postcss-scss": "4.0.3",
-    "prettier": "2.3.0",
+    "prettier": "3.3.3",
     "sass": "1.34.0",
     "sass-loader": "11.1.1",
     "snarkdown": "2.0.0",
